+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 # 0.0.4
-- 
+- Preserve entry list scroll position when adding, deleting, or reordering entries